requirements.txt
langchain
langchain_community
langchain-huggingface
wikipedia
python-dotenv
tiktoken
sentence-transformers
langchain-google-genai

